[CHORE] 패키지 및 줄 정리
