+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.INSOPTAndroidPractice"
         tools:targetApi="31">
         <activity
-            android:name="com.sopt.androidpractice.LoginActivity"
+            android:name=".presentation.login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -26,10 +26,10 @@
         </activity>
 
         <activity
-            android:name="com.sopt.androidpractice.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="false" />
         <activity
-            android:name="com.sopt.androidpractice.SignUpActivity"
+            android:name=".presentation.signup.SignUpActivity"
             android:exported="false" />
     </application>
