+++ app/src/main/java/org/sopt/sample/data/entity/request/RequestLoginDTO.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.data
+package org.sopt.sample.data.entity.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
