+++ app/src/main/java/org/sopt/sample/data/entity/response/ResponseLoginDTO.kt
@@ -1,16 +1,16 @@
-package org.sopt.sample.data
+package org.sopt.sample.data.entity.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseSignUpDTO(
+data class ResponseLoginDTO(
     @SerialName("status")
     val status: Int,
     @SerialName("message")
     val message: String,
-    @SerialName("newUser")
-    val newUser: User
+    @SerialName("result")
+    val result: User
 ) {
     @Serializable
     data class User(
