+++ app/src/main/java/org/sopt/sample/data/entity/response/ResponseSignUpDTO.kt
@@ -1,16 +1,16 @@
-package org.sopt.sample.data
+package org.sopt.sample.data.entity.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseLoginDTO(
+data class ResponseSignUpDTO(
     @SerialName("status")
     val status: Int,
     @SerialName("message")
     val message: String,
-    @SerialName("result")
-    val result: User
+    @SerialName("newUser")
+    val newUser: User
 ) {
     @Serializable
     data class User(
