+++ app/src/main/java/org/sopt/sample/data/service/UserInfoService.kt
@@ -1,5 +1,6 @@
-package org.sopt.sample.data
+package org.sopt.sample.data.service
 
+import org.sopt.sample.data.entity.response.ResponseUserInfoDTO
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
