+++ app/src/main/java/org/sopt/sample/presentation/home/Repo.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.home.data
+package org.sopt.sample.presentation.home
 
 data class Repo(
     val image: Int,
