+++ app/src/main/java/org/sopt/sample/presentation/login/LoginActivity.kt
@@ -1,19 +1,18 @@
-package com.sopt.androidpractice
+package org.sopt.sample.presentation.login
 
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import android.view.View
 import android.widget.Toast
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.widget.doAfterTextChanged
 import com.google.android.material.snackbar.Snackbar
-import org.sopt.sample.LoginViewModel
 import org.sopt.sample.R
 import org.sopt.sample.databinding.ActivityLoginBinding
+import org.sopt.sample.presentation.main.MainActivity
+import org.sopt.sample.presentation.signup.SignUpActivity
 
 class LoginActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginBinding
@@ -32,41 +31,6 @@ class LoginActivity : AppCompatActivity() {
         clickLogin()
         clickSignUp()
         setResultSignUp()
-
-        with(binding) {
-            etId.doAfterTextChanged { text ->
-                id = text.toString()
-                viewModel.checkPattern(text.toString(), 0)
-            }
-
-            etPassword.doAfterTextChanged { text ->
-                password = text.toString()
-                viewModel.checkPattern(text.toString(), 1)
-            }
-
-            viewModel.checkId.observe(this@LoginActivity) { check ->
-                if (check) {
-                    tvIdErrorMsg.visibility = View.GONE
-                } else {
-                    tvIdErrorMsg.visibility = View.VISIBLE
-                }
-                setLoginBtn()
-            }
-
-            viewModel.checkId.observe(this@LoginActivity) { check ->
-                if (check) {
-                    tvPasswordErrorMsg.visibility = View.GONE
-                } else {
-                    tvPasswordErrorMsg.visibility = View.VISIBLE
-                }
-                setLoginBtn()
-            }
-        }
-    }
-
-    private fun setLoginBtn() {
-        binding.btnLogin.isEnabled =
-            viewModel.checkId.value == true && viewModel.checkPassword.value == true
     }
 
     private fun clickLogin() {
@@ -76,7 +40,7 @@ class LoginActivity : AppCompatActivity() {
                 if (success) {
                     Toast.makeText(this@LoginActivity, R.string.login_success, Toast.LENGTH_SHORT)
                         .show()
-                    startActivity(intent)
+                    startActivity(Intent(this, MainActivity::class.java))
                 }
             }
         }
@@ -91,7 +55,6 @@ class LoginActivity : AppCompatActivity() {
 
     private fun setResultSignUp() {
         resultLauncher =
-            // getContent or PickVisualMedia 넣기 ?
             registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
                 if (result.resultCode == Activity.RESULT_OK) {
                     id = result.data?.getStringExtra(R.string.id.toString()) ?: ""
