+++ app/src/main/java/org/sopt/sample/presentation/login/LoginViewModel.kt
@@ -1,11 +1,11 @@
-package org.sopt.sample
+package org.sopt.sample.presentation.login
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import org.sopt.sample.data.RequestLoginDTO
-import org.sopt.sample.data.ResponseLoginDTO
-import org.sopt.sample.data.ServicePool.authService
+import org.sopt.sample.util.ServicePool.authService
+import org.sopt.sample.data.entity.request.RequestLoginDTO
+import org.sopt.sample.data.entity.response.ResponseLoginDTO
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -16,12 +16,6 @@ class LoginViewModel : ViewModel() {
 
     private val _successLogin = MutableLiveData<Boolean>()
     val successLogin: LiveData<Boolean> = _successLogin
-
-    private val _checkId = MutableLiveData<Boolean>()
-    val checkId: LiveData<Boolean> = _checkId
-
-    private val _checkPassword = MutableLiveData<Boolean>()
-    val checkPassword: LiveData<Boolean> = _checkPassword
 
     fun login(email: String, password: String) {
         authService.login(
@@ -42,13 +36,5 @@ class LoginViewModel : ViewModel() {
             override fun onFailure(call: Call<ResponseLoginDTO>, t: Throwable) {
             }
         })
-    }
-
-    fun checkPattern(text: String, type: Int): Boolean { // type 0 = id, 1 = password
-        when (type) {
-            0 -> _checkId.value = text.length in 6..10 || text.isNullOrBlank()
-            1 -> _checkPassword.value = text.length in 6..12 || text.isNullOrBlank()
-        }
-        return false
     }
 }
